Add Makefile for building/testing/linting project
Adds a Makefile that provides a build, clean, test, and lint command.
These commands are used for local development while goreleaser is still
used for release the go artifacts/docker container. The goreleaser
pipeline now leverages the make test command for running tests as a pre
hook to the release pipeline